requirements.txt
flask
flask-cors
gunicorn
sentence-transformers
chromadb
nltk
pytesseract
requests
beautifulsoup4
Pillow
